+++ backend/packages/eks/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @cat-factory/eks
 
+## 0.1.288
+
+### Patch Changes
+
+- Updated dependencies [1fbd83c]
+- Updated dependencies [00228c6]
+  - @cat-factory/contracts@0.287.1
+  - @cat-factory/kernel@0.281.2
+  - @cat-factory/integrations@0.153.7
+
+## 0.1.287
+
+### Patch Changes
+
+- Updated dependencies [bf473bd]
+  - @cat-factory/contracts@0.287.0
+  - @cat-factory/integrations@0.153.6
+  - @cat-factory/kernel@0.281.1
+
 ## 0.1.286
 
 ### Patch Changes