+++ backend/packages/gitlab/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @cat-factory/gitlab
 
+## 0.19.7
+
+### Patch Changes
+
+- Updated dependencies [1fbd83c]
+- Updated dependencies [00228c6]
+  - @cat-factory/contracts@0.287.1
+  - @cat-factory/kernel@0.281.2
+
+## 0.19.6
+
+### Patch Changes
+
+- Updated dependencies [bf473bd]
+  - @cat-factory/contracts@0.287.0
+  - @cat-factory/kernel@0.281.1
+
 ## 0.19.5
 
 ### Patch Changes