+++ frontend/app/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @cat-factory/app
 
+## 0.258.2
+
+### Patch Changes
+
+- Updated dependencies [1fbd83c]
+  - @cat-factory/contracts@0.287.1
+
+## 0.258.1
+
+### Patch Changes
+
+- Updated dependencies [bf473bd]
+  - @cat-factory/contracts@0.287.0
+
 ## 0.258.0
 
 ### Minor Changes